requirements.txt
torch
torchvision
transformers
roboflow
opencv-python
numpy
matplotlib
scikit-learn
Pillow
tqdm
pycocotools
